Delete variants and pattern matching.
I don’t yet have a compreshensive design for the syntax of variants and pattern matching. I also don’t  have a comprehensive design for the syntax of effects which are very closely related to variants. I’m gonna push forward with a feature set that does not include variants for now and circle back to include them later.